+++ mainProcess/src/controllers/appInfoReader.ts
@@ -1,6 +1,5 @@
 import { app } from "electron";
 
-
 export class AppInfoReader {
   private static get backstop() {
     const backstop = require("backstopjs");