+++ mainProcess/src/eventBuses/templateEventBus.ts
@@ -1,5 +1,5 @@
 import { ipcMain } from "electron";
-import { eventNames } from "../shared/constants/eventNames";
+import { eventNames } from "../../../shared/constants/eventNames";
 import { TemplateManager } from "../controllers/templateManager";
 
 ipcMain.handle(eventNames.CREATE_SCRIPT_TEMPLATE, (event, name: string, content: string) => {
@@ -16,4 +16,20 @@ ipcMain.handle(eventNames.RETRIEVE_SCRIPT_TEMPLATES, (event) => {
 
 ipcMain.handle(eventNames.DELETE_SCRIPT_TEMPLATE, (event, name: string) => {
   return TemplateManager.getInstance().deleteScriptTemplate(name);
+});
+
+ipcMain.handle(eventNames.CREATE_CONFIGURATION_TEMPLATE, (event, name: string, content: string) => {
+  return TemplateManager.getInstance().createConfigurationTemplate(name, content);
+});
+
+ipcMain.handle(eventNames.CREATE_OR_UPDATE_CONFIGURATION_TEMPLATE, (event, name: string, content: string) => {
+  return TemplateManager.getInstance().createOrUpdateConfigurationTemplate(name, content);
+});
+
+ipcMain.handle(eventNames.RETRIEVE_CONFIGURATION_TEMPLATES, () => {
+  return TemplateManager.getInstance().retrieveConfigurationTemplates();
+});
+
+ipcMain.handle(eventNames.DELETE_CONFIGURATION_TEMPLATE, (event, name: string) => {
+  return TemplateManager.getInstance().deleteConfigurationTemplate(name);
 });