+++ mainProcess/src/preload.ts
@@ -1,5 +1,8 @@
 import { contextBridge, ipcRenderer } from "electron";
+import { Titlebar, Color } from 'custom-electron-titlebar';
+import logger from "electron-log";
 
+let titleBar: Titlebar | null = null;
 
 declare global {
   interface Window {
@@ -9,6 +12,14 @@ declare global {
       receive: (channel: string, callback: (...args: any[]) => void) => void;
       receiveOnce: (channel: string, callback: (...args: any[]) => void) => void;
       invoke: (channel: string, ...args: any[]) => Promise<any>;
+      createTitleBar: () => void;
+      updateTitleBarTitle: (newTitle: string) => void;
+      logger: {
+        silly(...args: string[]): void;
+        info(...args: string[]): void;
+        warn(...args: string[]): void;
+        error(...args: string[]): void;
+      };
     };
   }
 }
@@ -32,6 +43,40 @@ const ipcHandler = {
 
   invoke: (channel: string, ...args: any[]) => {
     return ipcRenderer.invoke(channel, ...args);
+  },
+
+  createTitleBar: () => {
+    titleBar = new Titlebar({
+      backgroundColor: Color.fromHex('#02468a'),
+      icon: "../../assets/icon.png",
+    });
+    titleBar.updateTitle("Lemu");
+  },
+
+  updateTitleBarTitle: (newTitle: string) => {
+    if (titleBar) {
+      if (newTitle) {
+        newTitle = `Lemu - ${newTitle}`;
+      } else {
+        newTitle = "Lemu";
+      }
+      titleBar.updateTitle(newTitle);
+    }
+  },
+
+  logger: {
+    silly(...args: string[]): void {
+      logger.silly(...args);
+    },
+    info(...args: string[]): void {
+      logger.info(...args);
+    },
+    warn(...args: string[]): void {
+      logger.warn(...args);
+    },
+    error(...args: string[]): void {
+      logger.error(...args);
+    }
   }
 };