+++ src/components/templates/configuration/TemplateConfigurationWelcomeComponent.vue
@@ -1,0 +1,10 @@
+<template>
+  <v-row align="center" justify="center" class="ma-0">
+    <v-card>
+      <v-card-title>Configuration templates</v-card-title>
+      <v-card-text>
+        <p>Select a configuration template to edit it.</p>
+      </v-card-text>
+    </v-card>
+  </v-row>
+</template>