+++ src/store/applicationStore.ts
@@ -12,6 +12,18 @@ export default class ApplicationStore extends VuexModule {
   public snackbarSuccess: boolean = false;
   public appInfos: ApplicationInfo | null = null;
 
+  public get applicationTitle() {
+    const configName = this.context.rootGetters["configurationStore/configName"];
+    if (configName) {
+      const modified = this.context.rootGetters["configurationStore/hasConfigurationBeenModified"] ||
+        this.context.rootState.configurationStore.scriptsModified;
+
+      const saveText = modified ? "Unsaved" : "Saved";
+      return `${configName} - ${saveText}`;
+    }
+    return "";
+  }
+
   @Mutation
   public setSnackbarDisplayed({text, success}: {text: string, success: boolean}) {
     this.snackbarDisplayed = false;